perf: increase speed of Fast tokenizer
Around 2-3x faster in the larger benchmarks, no measurable difference
for the smaller benchmarks.

Comes at the cost of having to know the whole string beforehand (which
shouldn't be an issue when used through the Tokenizer).

Returns one-character slices into the input string instead of creating
a new string for each character.